requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery
redis
python-jose[cryptography]
passlib[bcrypt]
bcrypt
cryptography
google-auth
google-auth-oauthlib
google-api-python-client
msal
requests
beautifulsoup4
python-dotenv
python-multipart
PyPDF2
python-docx
zipfile36
